+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.2.0
 - AMaViS configuration is now possible using environment variables
 - Now all ClamAV logs are redirected as intended
-- Using alpine:latest since bug [9987](https://bugs.alpinelinux.org/issues/9987) was resolved
+- Supports SMTP client SASL authentication using Dovecot
 
 # 1.1.1
 - Make sure the .env settings are honored also for MYSQL